+++ Controllers/Freelancer/FreelancerLanguageController.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNetCore.Mvc;
 using MyCarearApi.Dtos;
 using MyCarearApi.Models;
 using MyCarearApi.Services;
+using MyCarearApi.Validations;
 
 namespace MyCarearApi.Controllers;
 
@@ -47,8 +48,10 @@ public class FreelancerLanguageController : ControllerBase
     {
         try
         {
-            if (!ModelState.IsValid)
-                return BadRequest();
+            var validate = new LanguageDtoValidation().Validate(language);
+            
+            if (!validate.IsValid)
+                return BadRequest(validate.Errors);
 
             string? userId = User.FindFirst(ClaimTypes.NameIdentifier) == null ? null : User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
 
@@ -72,6 +75,9 @@ public class FreelancerLanguageController : ControllerBase
     {
         try
         {
+            if (id < 0)
+                return BadRequest(error:"Id invalid");
+
             var result = await _languageService.Delete(id);
 
             if (!result.IsSuccess)
